# docs/swagger.yaml
basePath: /
definitions:
  banner_model.Banner:
    properties:
      banner_id:
        type: integer
      content:
        type: object
      created_at:
        type: string
      feature_id:
        type: integer
      is_active:
        type: boolean
      tag_ids:
        items:
          type: integer
        type: array
      updated_at:
        type: string
    type: object
  banner_model.BannerInsert:
    properties:
      content:
        type: object
      feature_id:
        type: integer
      is_active:
        type: boolean
      tag_id:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - content
    - feature_id
    - tag_id
    type: object
  banner_model.BannerUpdate:
    properties:
      banner_id:
        type: integer
      content:
        type: object
      feature_id:
        type: integer
      is_active:
        type: boolean
      tag_id:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - banner_id
    type: object
  transport.RespWriterBannerCreated:
    properties:
      banner_id:
        type: integer
    type: object
  transport.RespWriterError:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API сервер для сервиса баннеров
  title: Сервис баннеров
paths:
  /banner:
    delete:
      description: Удаления баннеров по фиче или тегу
      operationId: delete-banner-tag-feature
      parameters:
      - description: tag_id
        in: query
        name: tag_id
        type: integer
      - description: feature_id
        in: query
        name: feature_id
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
      security:
      - ApiKeyAuth: []
      summary: DeleteBannerOnTagOrFeature
      tags:
      - banner
    get:
      description: Получение всех баннеров c фильтрацией по фиче и/или тегу
      operationId: get-banner
      parameters:
      - description: tag_id
        in: query
        name: tag_id
        type: integer
      - description: feature_id
        in: query
        name: feature_id
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/banner_model.Banner'
            type: array
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.RespWriterError'
      security:
      - ApiKeyAuth: []
      summary: GetBanners
      tags:
      - banner
    post:
      consumes:
      - application/json
      description: Создание нового баннера
      operationId: create-banner
      parameters:
      - description: banner info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/banner_model.BannerInsert'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/transport.RespWriterBannerCreated'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.RespWriterError'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.RespWriterError'
      security:
      - ApiKeyAuth: []
      summary: CreateBanner
      tags:
      - banner
  /banner/{id}:
    delete:
      description: Удаление баннера по идентификатору
      operationId: delete-banner
      parameters:
      - description: id
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.RespWriterError'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.RespWriterError'
      security:
      - ApiKeyAuth: []
      summary: DeleteBanner
      tags:
      - banner
    patch:
      description: Обновление содержимого баннера
      operationId: update-banner
      parameters:
      - description: id
        in: path
        name: id
        type: integer
      - description: banner info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/banner_model.BannerUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.RespWriterError'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.RespWriterError'
      security:
      - ApiKeyAuth: []
      summary: UpdateBanner
      tags:
      - banner
  /user_banner:
    get:
      description: Получение баннера для пользователя
      operationId: get-user-banner
      parameters:
      - description: tag_id
        in: query
        name: tag_id
        required: true
        type: integer
      - description: feature_id
        in: query
        name: feature_id
        required: true
        type: integer
      - description: use_last_revision
        in: query
        name: use_last_revision
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.RespWriterError'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.RespWriterError'
      security:
      - ApiKeyAuth: []
      summary: GetUserBanner
      tags:
      - banner
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"